pyproject.toml
[tool.poetry]
name = "python-app"
version = "0.1.0"
description = "Aplicação Python com OpenAI"
authors = ["Pedro Ayres <pedroayres@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
openai = "^1.6.0"
uvicorn = "^0.23.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 